Fix compile errors

+++ src/bitlash-taskmgr.c
@@ -119,12 +119,12 @@ byte slot;
 	long next_wake_time = millis() + 500L;
 	for (slot=0; slot<NUMTASKS; slot++) {
 		if (tasklist[slot] != SLOT_FREE) {
-			if (waketime[slot] < next_wake_time) next_wake_time = waketime[slot];
+			if ((long)waketime[slot] < next_wake_time) next_wake_time = waketime[slot];
 		}
 	}
 	long millis_to_wait = next_wake_time - millis();
 	if (millis_to_wait < 0) millis_to_wait = 0;
-	return millis_to_wait;			// millis until next task runs
+	return (unsigned long)millis_to_wait;			// millis until next task runs
 }
 
 void showTaskList(void) {
